Added Repository and connect to JPA

+++ src/main/java/com/parkingSlot/Exception/ParkingSlotNotFoundException.java
@@ -1,7 +1,7 @@
 package com.parkingSlot.Exception;
 
 public class ParkingSlotNotFoundException extends RuntimeException {
-    public ParkingSlotNotFoundException(Long id) {
+    public ParkingSlotNotFoundException(int id) {
         super("Parking slot with id " + id + " not found");
     }
 }
